Moved test helpers, fixtures and matchers to their own packages.
